+++ ntfsprogs/ntfstruncate.c
@@ -270,7 +270,7 @@ static void parse_options(int argc, char *argv[])
 	if (*s2 || ll < 0 || (ll >= LLONG_MAX && errno == ERANGE))
 		err_exit("Invalid new length: %s\n", s);
 	new_len = ll;
-	ntfs_log_verbose("new length = %lli\n", new_len);
+	ntfs_log_verbose("new length = %lli\n", (long long)new_len);
 }
 
 /**
